requirements.txt
aniso8601==1.2.0
apipkg==1.4
asteval==0.9.1
autovizwidget==0.8.4
backports-abc==0.5
backports.shutil-get-terminal-size==1.0.0
backports.ssl-match-hostname==3.5.0.1
blist==1.3.6
certifi==2016.9.26
click==6.6
configparser==3.5.0
cov-core==1.15.0
coverage==3.7.1
DBUtils==1.1
decorator==4.0.10
entrypoints==0.2.2
enum34==1.1.6
execnet==1.4.1
findspark==1.1.0
Flask==0.12
Flask-RESTful==0.3.5
functools32==3.2.3.post2
futures==3.0.5
itsdangerous==0.24
Jinja2==2.8
jsonschema==2.5.1
kafka-python==1.3.1
MarkupSafe==0.23
mistune==0.7.3
mock==1.0.1
nbformat==4.2.0
nose==1.3.7
numpy==1.11.2
nupic==0.5.7
nupic.bindings==0.4.13
ordereddict==1.1
pandas==0.19.1
pathlib2==2.1.0
pexpect==4.2.1
pickleshare==0.7.4
plotly==1.10.0
prettytable==0.7.2
prompt-toolkit==1.0.9
psutil==1.0.1
ptyprocess==0.5.1
py4j
py==1.4.31
pycapnp==0.5.8
Pygments==2.1.3
pyproj==1.9.3
pytest==2.5.1
pytest-cov==1.6
pytest-xdist==1.8
python-dateutil==2.1
pytz==2016.10
PyYAML==3.10
pyzmq==16.0.2
qtconsole==4.2.1
requests==2.12.3
scipy==0.18.1
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.10.0
sparkmagic==0.8.4
terminado==0.6
tornado==4.4.2
traitlets==4.3.1
unittest2==0.5.1
uWSGI==2.0.14
validictory==0.9.1
wcwidth==0.1.7
Werkzeug==0.11.15
